refactor(dashboards): move TreemapChart into the treemap visualization dir
The dashboard visualization is the chart's primary owner; test-performance
imports it from there instead of a top-level shared component.
